Refactor agent interface to add modularity and customization
Additions to the agent interface enable customizable settings, including
the ability to define the vector store, embedding model, and memory.
These changes support modularity and flexibility for different use
cases. It also simplifies the creation and fetching of agent settings.
This introduces a more adaptable and user-friendly system for creating
agent instances. The commit also includes refactoring and removing of
commented code and unused imports.

Closes #123, #456
